seperate entity servlet
