+++ byway/src/main/java/com/google/sps/data/LoginState.java
@@ -17,11 +17,11 @@ package com.google.sps.data;
 import com.google.errorprone.annotations.Immutable;
 
 @Immutable
-public final class Login {
+public final class LoginState {
   private final boolean status;
   private final String url;
 
-  public Login(boolean status, String url) {
+  public LoginState(boolean status, String url) {
     this.status = status;
     this.url = url;
   }
