+++ byway/src/main/java/com/google/sps/servlets/CreateTripServlet.java
@@ -20,31 +20,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/entity")
-public class EntityServlet extends HttpServlet {
+@WebServlet("/api/createtrip")
+public class CreateTripServlet extends HttpServlet {
   private final Gson gson = new Gson();
   private final UserService userService = UserServiceFactory.getUserService();
   private final DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-
-  @Override
-  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    com.google.appengine.api.users.User user = userService.getCurrentUser();
-    UserInfo userInfo = addUserEntity(user);
-    List<String> userTripIds = userInfo.getTripIds();
-    ArrayList<Trip> userTrips = new ArrayList<Trip>();
-    for (String tripId : userTripIds) {
-      try {
-        Entity tripEntity = datastore.get(KeyFactory.stringToKey(tripId));
-        Trip trip = Trip.fromEntity(tripEntity);
-        userTrips.add(trip);
-      } catch (EntityNotFoundException exception) {
-        return;
-      }
-    }
-    String json = gson.toJson(userTrips);
-    response.setContentType("application/json");
-    response.getWriter().println(json);
-  }
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
@@ -73,29 +53,6 @@ public class EntityServlet extends HttpServlet {
     tripEntity.setProperty("keyString", KeyFactory.keyToString(tripKey));
     datastore.put(tripEntity);
     return tripKey;
-  }
-
-  public UserInfo addUserEntity(com.google.appengine.api.users.User user) {
-    System.setProperty(
-        DatastoreServiceConfig.DATASTORE_EMPTY_LIST_SUPPORT, Boolean.TRUE.toString());
-    // Create a key based on the user ID
-    Key userKey = KeyFactory.createKey(UserInfo.DATASTORE_ENTITY_KIND, user.getUserId());
-    UserInfo userInfo;
-    try {
-      // try to retrieve the entity with the key
-      Entity userEntity = datastore.get(userKey);
-      datastore.put(userEntity);
-      userInfo = UserInfo.fromEntity(userEntity);
-    } catch (EntityNotFoundException exception) {
-      // If the user doesn't exist yet or is new, create a new user
-      Entity newUserEntity = new Entity(userKey);
-      newUserEntity.setProperty("email", user.getEmail());
-      newUserEntity.setProperty("userId", user.getUserId());
-      newUserEntity.setProperty("tripIds", new ArrayList<String>());
-      datastore.put(newUserEntity);
-      userInfo = UserInfo.fromEntity(newUserEntity);
-    }
-    return userInfo;
   }
 
   public void addTripForUser(String tripKey) {
